These are the merged files 990 data and EOBMF data
These files are the merged datasets for the (now combined) Exempt
Organizations Business Master File Extract (2015) and the 2012 data of
the Internal Revenue Service collected  Annual Extract of Tax-Exempt
Organization Financial Data.  The three files represent the Form 990
Extract, Form 990-EZ Extract, and the Form 990-PF Extract combined with
the  EOBMF of all recorded regions, domestically and international.
Included also in the Extract Documentation which explains the variables
collected.